+++ components/Layout/Layout.js
@@ -1,0 +1,10 @@
+import Header from '../Header'
+
+const Layout = props => (
+  <>
+  <Header />
+  {props.children}
+  </>
+)
+
+export default Layout;
