--- pages/Auth/Index.js
@@ -1,1 +1,0 @@
-export { default } from './Auth';
