+++ pages/index.js
@@ -1,20 +1,27 @@
-import { GlobalStyle } from './GlobalStyle';
-
+import styled from 'styled-components';
 import Link from 'next/link';
-import Auth from './Auth'
+import Layout from '../components/Layout';
 
 function Index() {
   return (
-    <>
-    <GlobalStyle />
-      <div>
-        <p>index Page</p>
-        <Link href='/Auth'>
-          <a>Login</a>
-        </Link>
-      </div>
-    </>
+    <Layout>
+      <Box>
+      <video muted autoPlay loop style={{ width: `100%` }}>
+        <source src="../static/wework.mp4" type="video/mp4" />
+        Your browser does not support the video tag.
+      </video>
+      </Box>
+    </Layout>
   );
 }
+
+const Box = styled.div`
+  position: absolute;
+  top: 0;
+  left: 0;
+  height: 710px;
+  width: 100%;
+  overflow: hidden;
+`
 
 export default Index;
